# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ErrorVo": {
				"properties": {
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"messageCode": {
						"type": "string"
					},
					"events": {
						"items": {
							"$ref": "#/components/schemas/EventParamsType"
						},
						"type": "array"
					}
				},
				"required": [
					"code",
					"message",
					"messageCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonCodeGroupQueryResponse": {
				"properties": {
					"codeGroup": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"order": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"codeGroup",
					"name",
					"description",
					"order"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonResponse_CommonCodeGroupQueryResponse-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/CommonCodeGroupQueryResponse"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"CommonCode": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdId": {
						"type": "number",
						"format": "double"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"codeGroup": {
						"type": "string"
					},
					"code": {
						"type": "string"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"updatedId",
					"createdId",
					"order",
					"description",
					"name",
					"codeGroup",
					"code"
				],
				"type": "object",
				"description": "Model CommonCode\nThis model or at least one of its fields has comments in the database, and requires an additional setup for migrations: Read more: https://pris.ly/d/database-comments"
			},
			"CommonResponse_CommonCode-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/CommonCode"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"AttachFile": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdId": {
						"type": "number",
						"format": "double"
					},
					"fileSize": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"fileExt": {
						"type": "string",
						"nullable": true
					},
					"fileChgName": {
						"type": "string",
						"nullable": true
					},
					"fileName": {
						"type": "string"
					},
					"fileUrl": {
						"type": "string"
					},
					"path": {
						"type": "string",
						"nullable": true
					},
					"attachFileId": {
						"type": "string"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"updatedId",
					"createdId",
					"fileSize",
					"fileExt",
					"fileChgName",
					"fileName",
					"fileUrl",
					"path",
					"attachFileId"
				],
				"type": "object",
				"description": "Model AttachFile\nThis model or at least one of its fields has comments in the database, and requires an additional setup for migrations: Read more: https://pris.ly/d/database-comments"
			},
			"CommonResponse_AttachFile_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/AttachFile"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"SendCertificationSMSResponse": {
				"properties": {
					"messageId": {
						"type": "string",
						"description": "문자 전송 업체에서 반환하는 UUID"
					},
					"expiryTime": {
						"type": "number",
						"format": "double",
						"description": "만료시간 epoch time"
					}
				},
				"required": [
					"messageId",
					"expiryTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonResponse_SendCertificationSMSResponse_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/SendCertificationSMSResponse"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"SendCertificationSMSRequest": {
				"properties": {
					"consoleMemberName": {
						"type": "string",
						"description": "콘솔멤버 이름",
						"example": "김개발"
					},
					"nationCodeNumber": {
						"type": "string",
						"description": "국가번호",
						"example": "82"
					},
					"phone": {
						"type": "string",
						"description": "핸드폰 번호",
						"example": "01042128867"
					}
				},
				"required": [
					"consoleMemberName",
					"nationCodeNumber",
					"phone"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonResponse": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"CertificateCodeRequest": {
				"properties": {
					"code": {
						"type": "string",
						"description": "핸드폰 인증 인증번호 6자리",
						"minLength": 6,
						"maxLength": 6
					}
				},
				"required": [
					"code"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CommonResponse_App_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/App"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"CreateAppRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "앱 이름",
						"example": "appName",
						"minLength": 1,
						"maxLength": 50
					},
					"iconUrl": {
						"type": "string",
						"example": "https://example.com/example/exampleexample",
						"pattern": "[^(?:/(?:\\*{1,2}|\\w+\\*{0,2}))+/?$]"
					},
					"blockChainId": {
						"type": "string",
						"description": "체인",
						"example": "ETH"
					}
				},
				"required": [
					"name",
					"iconUrl",
					"blockChainId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Omit_CreateAppRequest.blockChainId_": {
				"$ref": "#/components/schemas/Pick_CreateAppRequest.Exclude_keyofCreateAppRequest.blockChainId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"UpdateAppRequest": {
				"$ref": "#/components/schemas/Omit_CreateAppRequest.blockChainId_"
			},
...
			"CommonResponse_AppDetailResponseType_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/AppDetailResponseType"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"FindManyAppsQueryResponse": {
				"properties": {
					"appId": {
						"type": "number",
						"format": "double",
						"description": "앱 id"
					},
					"name": {
						"type": "string",
						"description": "앱 이름"
					},
					"iconUrl": {
						"type": "string",
						"nullable": true,
						"description": "앱 아이콘 url"
					},
					"appWorkspace": {
						"items": {
							"properties": {
								"defaultYn": {
									"type": "string",
									"description": "디폴트 워크스페이스 여부"
								},
								"workspaceId": {
									"type": "number",
									"format": "double",
									"description": "워크스페이스 id"
								}
							},
							"required": [
								"defaultYn",
								"workspaceId"
							],
							"type": "object"
						},
						"type": "array",
						"description": "앱 하위 워크 스페이스 목록"
					},
					"appWallet": {}
				},
				"required": [
					"appId",
					"name",
					"iconUrl",
					"appWorkspace",
					"appWallet"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonResponse_FindManyAppsQueryResponse-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/FindManyAppsQueryResponse"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"CommonResponse_AppApiKey_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/AppApiKey"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"CreateAppApiKeyRequest": {
				"properties": {
					"appId": {
						"type": "integer",
						"format": "int32",
						"description": "앱 식별자",
						"example": 1
					},
					"name": {
						"type": "string",
						"description": "앱 API KEY 이름",
						"minLength": 1,
						"maxLength": 50
					},
					"apiKeyKindCode": {
						"type": "string",
						"description": "코드그룹명 : [API_KEY]\r\n키유형코드 =>\r\nSIGN_SECRET: 서명 시크릿\r\nREST_API: REST API 키\r\nOAUTH2_CLIENT: oauth2클라이언트\r\nENC_PUBLIC: 암호화 공개키\r\nENC_PRIVATE: 암호화 비공개키\r\nCLIENT_ID : 클라이언트 ID\r\nCLIENT_SECRET : 클라이언트 시크릿",
						"example": "REST_API"
					}
				},
				"required": [
					"appId",
					"name",
					"apiKeyKindCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommonResponse_AppApiKey-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/AppApiKey"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"CommonResponse_AppWalletBalance-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/AppWalletBalance"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"SupportChainId": {
				"enum": [
					137,
					80001,
					56,
					97
				],
				"type": "number"
			},
...
			"CommonResponse_AppWebhook_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"$ref": "#/components/schemas/AppWebhook"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"CreateAppWebhookRequest": {
				"properties": {
					"appId": {
						"type": "number",
						"format": "double",
						"description": "앱id"
					},
					"profile": {
						"type": "string",
						"description": "프로필 : DEV(개발환경,테스트) , PRD(운영)"
					},
					"name": {
						"type": "string",
						"description": "웹훅명",
						"minLength": 1,
						"maxLength": 20
					},
					"encryptionYn": {
						"type": "string",
						"description": "구간암호화 여부 Y/N"
					},
					"endpointUrl": {
						"type": "string",
						"description": "엔드포인트 URL"
					},
					"webhookId": {
						"type": "number",
						"format": "double",
						"description": "웹훅 id"
					}
				},
				"required": [
					"appId",
					"profile",
					"name",
					"encryptionYn",
					"endpointUrl",
					"webhookId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateAppWebhookRequest": {
				"$ref": "#/components/schemas/CreateAppWebhookRequest"
			},
			"CommonResponse_AppWebhook-Array_": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/AppWebhook"
						},
						"type": "array"
					},
					"error": {
						"$ref": "#/components/schemas/ErrorVo"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
			"DeleteAppWebhookRequest": {
				"properties": {
					"appId": {
						"type": "number",
						"format": "double"
					},
					"profile": {
						"type": "string",
						"enum": [
							"DEV",
							"PRD"
						]
					},
					"webhookId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"appId",
					"profile",
					"webhookId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"devAuth": {
				"type": "http",
				"scheme": "basic"
			},
			"basic": {
				"type": "http",
				"scheme": "basic"
			},
			"token": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"cookie": {
				"type": "apiKey",
				"name": "accessToken",
				"in": "cookie"
			}
		}
	},
	"info": {
		"title": "platpome-console-backend",
		"version": "0.0.1",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/common/code-groups": {
			"get": {
				"operationId": "FindManyCodeGroups",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_CommonCodeGroupQueryResponse-Array_"
								}
							}
						}
					}
				},
				"description": "공통 코드 그룹 목록 조회 API",
				"tags": [
					"[공통] Common"
				],
				"security": [],
				"parameters": [
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/common/codes": {
			"get": {
				"operationId": "FindManyCodes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_CommonCode-Array_"
								}
							}
						}
					}
				},
				"description": "공통 코드 목록 조회 API",
				"tags": [
					"[공통] Common"
				],
				"security": [],
				"parameters": [
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "codeGroup",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/common/s3": {
			"post": {
				"operationId": "CreateAttachFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AttachFile_"
								}
							}
						}
					}
				},
				"description": "S3 파일 업로드 및 파일정보 DB 저장 API",
				"tags": [
					"[공통] Common"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/common/sms/certification": {
			"post": {
				"operationId": "SendCertificationSMS",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_SendCertificationSMSResponse_"
								}
							}
						}
					}
				},
				"description": "핸드폰 인증 > 문자 발송 API",
				"tags": [
					"[콘솔 회원] ConsoleMember",
					"[공통] Common"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendCertificationSMSRequest"
							}
						}
					}
				}
			}
		},
		"/common/sms/certification/{messageId}": {
			"post": {
				"operationId": "CertificateCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse"
								}
							}
						}
					}
				},
				"description": "핸드폰 인증 > 번호 인증 API",
				"tags": [
					"[콘솔 회원] ConsoleMember",
					"[공통] Common"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "messageId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CertificateCodeRequest"
							}
						}
					}
				}
			}
		},
...
		"/apps": {
			"post": {
				"operationId": "CreateApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_App_"
								}
							}
						}
					}
				},
				"description": "콘솔 메인 홈 > 앱 등록 API",
				"tags": [
					"[앱] App"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAppRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindManyApps",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_FindManyAppsQueryResponse-Array_"
								}
							}
						}
					}
				},
				"description": "콘솔 회원이 속한 앱 목록 조회 API",
				"tags": [
					"[앱] App"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "앱 이름 검색 값 (포함 검색)",
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/apps/{appId}": {
			"put": {
				"operationId": "UpdateApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_App_"
								}
							}
						}
					}
				},
				"description": "앱 정보 수정 API",
				"tags": [
					"[앱] App"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAppRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindApp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppDetailResponseType_"
								}
							}
						}
					}
				},
				"description": "앱 상세 조회 API",
				"tags": [
					"[앱] App"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"description": "= 앱 아이디",
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/apps/api-keys": {
			"post": {
				"operationId": "CreateAppApiKeys",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppApiKey_"
								}
							}
						}
					}
				},
				"description": "앱 API KEY 발급 API",
				"tags": [
					"[앱 API] AppApi",
					"[앱] App"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAppApiKeyRequest"
							}
						}
					}
				}
			}
		},
		"/apps/api-keys/{appId}": {
			"get": {
				"operationId": "FindManyAppApiKeys",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppApiKey-Array_"
								}
							}
						}
					}
				},
				"description": "앱 API KEY 목록 조회 API",
				"tags": [
					"[앱 API] AppApi",
					"[앱] App"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/apps/{appId}/chains/{chainId}/tokens": {
			"get": {
				"operationId": "GetAppWalletBalance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppWalletBalance-Array_"
								}
							}
						}
					}
				},
				"description": "앱 지갑 보유 토큰 조회",
				"tags": [
					"[앱] App"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "chainId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/SupportChainId"
						}
					},
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/apps/api-keys/{apiKeyId}": {
			"delete": {
				"operationId": "DeleteAppApiKey",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse"
								}
							}
						}
					}
				},
				"description": "앱 API KEY 삭제 API",
				"tags": [
					"[앱 API] AppApi",
					"[앱] App"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "apiKeyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/webhooks/apps": {
			"post": {
				"operationId": "CreateAppWebhook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppWebhook_"
								}
							}
						}
					}
				},
				"description": "앱 웹훅 등록 API",
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAppWebhookRequest"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateAppWebhook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppWebhook_"
								}
							}
						}
					}
				},
				"description": "앱 웹훅 수정 API",
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAppWebhookRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAppWebhook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppWebhook_"
								}
							}
						}
					}
				},
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteAppWebhookRequest"
							}
						}
					}
				}
			}
		},
		"/webhooks/apps/{appId}": {
			"get": {
				"operationId": "FindManyAppWebhooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommonResponse_AppWebhook-Array_"
								}
							}
						}
					}
				},
				"description": "앱 웹훅 목록 조회 API",
				"tags": [
					"[웹훅] Webhook"
				],
				"security": [
					{
						"token": []
					},
					{
						"cookie": []
					},
					{
						"devAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "appId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "가져올 열 수.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지 번호.",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "'DEV' | 'PRD'",
						"in": "query",
						"name": "profile",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"DEV",
								"PRD"
							]
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}